==> R/0-ConfigurationVariables.R <==
# Threshold on the number of minimum patients per gene
# 0 for all genes and patients
minNumOfPatientsPerGene <<- 30   #5 for ~500 patients and genes in BRCA
                                #20 -> 1456
minNumOfMutationsPerPatient <<- 5 #5
minNumOfPatientsPerClinical <<- 5 #5
